+++ packages/dlt-source-aquabyte/src/dlt_source_aquabyte/__init__.py
@@ -2,7 +2,8 @@
 
 from importlib.metadata import version
 
-from dlt_source_aquabyte.aquabyte import MAX_WINDOW_DAYS, aquabyte_source
+from dlt_source_aquabyte.aquabyte import aquabyte_source
+from dlt_source_aquabyte.windows import MAX_WINDOW_DAYS
 
 __version__ = version("dlt-source-aquabyte")
 __all__ = ["MAX_WINDOW_DAYS", "aquabyte_source"]